Refactor integration tests for clearer structure.
